update readme couple of level-related fixes fix some harmless warnings generate build-info file and put some compile information into it, + display in show version command.
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@7779 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ servmon/smalrm.cxx
@@ -51,7 +51,7 @@ mgr::hasalarm(str host, str metric)
 int
 mgr::low(int value)
 {
-	return value * 0.75;
+	return int(value * 0.75);
 }
 
 void
